% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_hhs_attainment_functions.R
\name{hhs_attainment}
\alias{hhs_attainment}
\title{Get clean attainment data.}
\usage{
hhs_attainment(academicYear, yearGroup)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}
}
\description{
Returns clean details of attainment in the chosen academic year for chosen
year group.
}
\examples{
hhs_attainment(2020, "10")
}
