% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_hhs_attainment_functions.R
\name{hhs_attainment_multiple}
\alias{hhs_attainment_multiple}
\title{Get clean attainment data for multiple year groups.}
\usage{
hhs_attainment_multiple(academicYear, yearGroupFrom = "7", yearGroupTo = "11")
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroupFrom}{starting year group as string.  Defaults to "7".}

\item{yearGroupTo}{ending year group as string.  Defaults to "11".}
}
\description{
Returns clean details of attainment in the chosen academic year for multiple
year groups.
}
\examples{
hhs_attainment_multiple(2021, "7", "11")
}
