% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hhs_attendance_functions.R
\name{hhs_attendance_student_lesson}
\alias{hhs_attendance_student_lesson}
\title{Get clean student lesson marks.}
\usage{
hhs_attendance_student_lesson(academicYear, goDate)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDate}{date as string in the form yyyy-mm-dd.}
}
\description{
Returns clean details of student lesson marks for chosen
academic year and date.
}
\examples{
hhs_attendance_student_lesson(2020, "2020-07-01")
}
