% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hhs_attendance_functions.R
\name{hhs_attendance_student_lesson_range}
\alias{hhs_attendance_student_lesson_range}
\title{Get clean student lesson marks over a range of dates.}
\usage{
hhs_attendance_student_lesson_range(academicYear, goDateStart, goDateEnd)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDateStart}{start date of range as string in the form yyyy-mm-dd.}

\item{goDateEnd}{end date of range as string in the form yyyy-mm-dd.}
}
\description{
Returns clean student lesson marks over a range of dates for chosen
academic year.
}
\examples{
hhs_attendance_student_lesson_range(2021, "2021-07-12", "2021-07-16")
}
