% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hhs_attendance_functions.R
\name{hhs_attendance_student_summary}
\alias{hhs_attendance_student_summary}
\title{Get clean student attendance summary.}
\usage{
hhs_attendance_student_summary(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean details of student attendance summary for chosen
academic year.
}
\examples{
hhs_attendance_student_summary(2020)
}
