% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_hhs_behaviour_functions.R
\name{hhs_behaviour_events_range}
\alias{hhs_behaviour_events_range}
\title{Get clean behaviour events.}
\usage{
hhs_behaviour_events_range(academicYear, goDateStart, goDateEnd)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDateStart}{start date of range as string in the form yyyy-mm-dd.}

\item{goDateEnd}{end date of range as string in the form yyyy-mm-dd.}
}
\description{
Returns clean behaviour events over a range of dates within in the chosen
academic year.
}
\examples{
hhs_behaviour_events_range(2021, "2021-07-12", "2021-07-16")
}
