% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_hhs_class_list_functions.R
\name{hhs_class_list_student}
\alias{hhs_class_list_student}
\title{Get student-focused class list.}
\usage{
hhs_class_list_student(academicYear, student)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{student}{student search string.}
}
\description{
Returns clean details of all classes in the chosen academic year for chosen
student.
}
\examples{
hhs_class_list_student(2020, "smith")
}
