% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_attainment_functions.R
\name{hhs_exam_results}
\alias{hhs_exam_results}
\title{Get clean external examination results.}
\usage{
hhs_exam_results(academicYear, yearGroup, type = NULL)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}

\item{type}{qualification type as string.}
}
\description{
Returns clean details of external examination results in the chosen
academic year for chosen
year group.
}
\examples{
hhs_exam_results(2020, "11")
hhs_exam_results(2020, "11", type = "gcse")
hhs_exam_results(2020, "11", type = c("gcse", "btec"))
}
