% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_attainment_functions.R
\name{hhs_markbook_attainment}
\alias{hhs_markbook_attainment}
\title{Get clean markbook attainment.}
\usage{
hhs_markbook_attainment(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean details of attainment in the chosen academic year for all
markbooks.
}
\examples{
hhs_markbook_attainment(2022)
}
