% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_student_details_functions.R
\name{hhs_raw_student_details_general}
\alias{hhs_raw_student_details_general}
\title{Get raw general student details.}
\usage{
hhs_raw_student_details_general(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns raw general student details for chosen academic year.
}
\examples{
hhs_student_details_general(2021)
}
