% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_hhs_readability_functions.R
\name{hhs_readability}
\alias{hhs_readability}
\title{Get clean readability data.}
\usage{
hhs_readability(text_dir, return_full = FALSE)
}
\arguments{
\item{text_dir}{path to directory containing texts as string.}

\item{return_full}{return full readability statistics as logical.  Default FALSE.}
}
\description{
Returns clean readability data for texts in a given directory.
}
\examples{
hhs_readability("C:/Users//User/Texts", TRUE)
}
