% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hhs_school_functions.R
\name{hhs_school_population}
\alias{hhs_school_population}
\title{Get clean school population data.}
\usage{
hhs_school_population()
}
\description{
Returns clean details of school population for all available academic years,
along with a line plot.
}
\examples{
hhs_school_population()
}
