% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hhs_setup_functions.R
\name{hhs_setup}
\alias{hhs_setup}
\title{Run set up for gfs functions.}
\usage{
hhs_setup(api_key = NULL)
}
\arguments{
\item{api_key}{API key as string.}
}
\description{
Use this function to store your API key for use with hhs functions.
Authentication is required for all requests.  See your school's
administrator for an API key.
}
\examples{
hhs_setup("a96650481f0245eea396726f85ac7049")
}
