% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_student_details_functions.R
\name{hhs_student_send_search}
\alias{hhs_student_send_search}
\title{Get clean SEND notes search.}
\usage{
hhs_student_send_search(academicYear, notesSearch)
}
\arguments{
\item{academicYear}{notesSearch as string.}
}
\description{
Returns clean student details based on SEND notes search for chosen
academic year.
}
\examples{
hhs_student_send_search(2021, "extra time")
hhs_student_send_search(2021, "irlen")
}
