% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_attainment_functions.R
\name{hhs_targets}
\alias{hhs_targets}
\title{Get clean targets.}
\usage{
hhs_targets(academicYear, subject = NULL)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{subject}{subject as string.  .  Defaults to null.}
}
\description{
Returns clean details of targets in the chosen academic year and subject for all
year groups.
}
\examples{
hhs_targets(2022, "Combined Science 9-1")
}
