% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_hhs_timetable_functions.R
\name{hhs_timetable}
\alias{hhs_timetable}
\title{Get clean timetable data.}
\usage{
hhs_timetable(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean timetable data for chosen academic year.
}
\examples{
hhs_timetable(2021)
}
