% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hhs_behaviour_functions.R
\name{hhs_exclusions}
\alias{hhs_exclusions}
\title{Get clean school exclusion data.}
\usage{
hhs_exclusions(academicYear, goDateStart, goDateEnd)
}
\description{
Returns clean school exclusion data for chosen academic year and date range,
along with a line plot.
}
\examples{
hhs_exclusions(2022, "2021-09-01", "2021-09-30")
}
