% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hhs_attainment_functions.R
\name{hhs_targets_science}
\alias{hhs_targets_science}
\title{Get clean science targets.}
\usage{
hhs_targets_science(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean details of science targets in the chosen academic year for all
year groups.
}
\examples{
hhs_targets_science(2022)
}
